+++ src/ImageRenderer.c
@@ -42,8 +42,11 @@ static void image_freeData(WINFO* p) {
 	p->r_data = NULL;
 }
 
-static int image_supportsMode(int mode) {
-	return (mode & (SHOW_CARET_LINE_HIGHLIGHT | SHOW_CONTROL_CHARS | SHOW_LINENUMBERS | SHOW_CARET_PRESERVE_COLUMN | SHOW_SYNTAX_HIGHLIGHT)) == 0;
+static int image_supportsMode(EDIT_MODE mode) {
+	if (!mode.em_displayMode) {
+		return FALSE;
+	}
+	return (mode.em_flag & (SHOW_CARET_LINE_HIGHLIGHT | SHOW_CONTROL_CHARS | SHOW_LINENUMBERS | SHOW_CARET_PRESERVE_COLUMN | SHOW_SYNTAX_HIGHLIGHT)) == 0;
 }
 
 static HANDLE image_createHandle(FTABLE* fp) {